backend/requirements.txt
# Core
fastapi>=0.115.6
uvicorn[standard]==0.34.0
websockets>=14.1
python-dotenv>=1.0.1
pydantic>=2.10.4

# AI
google-genai>=1.0.0
browser-use>=0.12.0
langchain-google-genai>=4.0.0

# Voice
faster-whisper==1.1.0
sounddevice==0.5.1
numpy==2.2.1

# TTS
httpx>=0.28.1
aiofiles==24.1.0

# Tools
mss==9.0.2
Pillow>=11.1.0
pyautogui==0.9.54
playwright>=1.49.1

# Google APIs
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.159.0

# Telegram
python-telegram-bot>=20.0

# System stats (for /status command)
psutil>=5.9.0
